saving data to database
