+++ package.json
@@ -14,6 +14,7 @@
     "react-redux": "^8.0.4",
     "react-router-dom": "^6.4.2",
     "react-scripts": "5.0.1",
+    "react-loading-icons": "^1.1.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
